# api/openapi.yaml
openapi: 3.1.0
info:
  title: Raincheck API
  version: 0.1.0
  description: All calendar values use UTC. Raw participant edit tokens occur only in successful join responses.
paths:
...
  /api/groups/{inviteCode}/polls/{pollId}/responses/me:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }, { $ref: '#/components/parameters/PollId' }, { $ref: '#/components/parameters/ParticipantToken' }]
    post:
      summary: Create a draft response on an open poll
      responses: { '201': { $ref: '#/components/responses/Response' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
    get:
      summary: Read the authenticated participant response
      responses: { '200': { $ref: '#/components/responses/Response' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' } }
    put:
      summary: Replace intervals on an open poll; identical payload preserves IDs and confirmation
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [intervals], properties: { intervals: { type: array, items: { $ref: '#/components/schemas/IntervalInput' } } } } } } }
      responses: { '200': { $ref: '#/components/responses/Response' }, '400': { $ref: '#/components/responses/InvalidRequest' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' }, '422': { $ref: '#/components/responses/InvalidSchedule' } }
    delete:
      summary: Delete the response from an open poll; closed-poll responses are immutable
      responses: { '204': { description: Deleted }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
  /api/groups/{inviteCode}/polls/{pollId}/responses/me/confirm:
    parameters: [{ $ref: '#/components/parameters/InviteCode' }, { $ref: '#/components/parameters/PollId' }, { $ref: '#/components/parameters/ParticipantToken' }]
    post:
      summary: Confirm a response on an open poll idempotently
      responses: { '200': { $ref: '#/components/responses/Response' }, '401': { $ref: '#/components/responses/Error' }, '404': { $ref: '#/components/responses/Error' }, '409': { $ref: '#/components/responses/Error' } }
...
components:
  parameters:
    InviteCode: { name: inviteCode, in: path, required: true, schema: { type: string, pattern: '^[A-Za-z0-9_-]{32,64}$' } }
    PollId: { name: pollId, in: path, required: true, schema: { type: string, format: uuid } }
    ParticipantToken: { name: X-Participant-Token, in: header, required: true, schema: { type: string, minLength: 32, maxLength: 128 } }
...
  responses:
    Error: { description: Stable API error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    InvalidRequest: { description: INVALID_REQUEST, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    InvalidSchedule: { description: INVALID_SCHEDULE, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
...
    Response: { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Response' } } } }
...
  schemas:
    Error: { type: object, required: [error], properties: { error: { type: object, required: [code, message, requestId], properties: { code: { enum: [INVALID_REQUEST, UNAUTHORIZED, GROUP_NOT_FOUND, POLL_NOT_FOUND, RESPONSE_NOT_FOUND, PARTICIPANT_NAME_TAKEN, RESPONSE_ALREADY_EXISTS, POLL_STATE_CONFLICT, INVALID_SCHEDULE, INTERNAL_ERROR] }, message: { type: string }, requestId: { type: string } } } } }
...
    IntervalInput: { type: object, required: [localDate, startTime, endTime, kind, preferenceDirection], properties: { localDate: { type: string, format: date }, startTime: { type: string }, endTime: { type: string }, kind: { enum: [UNAVAILABLE, IF_NEEDED, PREFERRED] }, preferenceDirection: { type: [string, 'null'], enum: [EARLIER, FLAT, LATER, null] } } }
    Interval: { allOf: [{ $ref: '#/components/schemas/IntervalInput' }, { type: object, required: [id], properties: { id: { type: string, format: uuid } } }] }
    Response: { type: object, required: [id, state, confirmedAt, intervals], properties: { id: { type: string, format: uuid }, state: { enum: [DRAFT, CONFIRMED] }, confirmedAt: { type: [string, 'null'], format: date-time }, intervals: { type: array, items: { $ref: '#/components/schemas/Interval' } } } }
...